+++ src/main/java/com/navangs/maribong/repository/QuizRepository.java
@@ -3,6 +3,6 @@ package com.navangs.maribong.repository;
 import com.navangs.maribong.dao.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<Quiz, Long> {
-    Quiz findQuizById(Long id);
+public interface QuizRepository extends JpaRepository<Quiz, Integer> {
+    Quiz findQuizById(Integer id);
 }
